:recycle: Prepare SwiftBuild
